+++ ADMF.Core/changelog.md
@@ -2,6 +2,8 @@
 
 ## ???
 
+- New: Command Sync-AdcObject - Replicate a single item between DCs.
+- New: Command Get-AdcExchangeVersion - Get a list of exchange versions and the various identifiers to see which has been deployed.
 - Upd: Incremented PSFramework minimum version.
 
 ## 1.0.0 (2020-09-10)